create SQLAlchemy class
